+++ CHANGELOG.md
@@ -3,8 +3,9 @@
 ## Breaking Changes
 
 - Vue 2 support has been dropped.
-- The CommonJS build of coalesce-vue has been dropped.
+- The CommonJS build of coalesce-vue has been dropped - only the ESM build remains. Most projects should be unaffected.
 - `ViewModel.$getErrors` now returns a `string[]` instead of a `Generator<string>`.
+- `StandardBehaviors.AfterDelete` is now `AfterDeleteAsync` and has a different signature and semantics. Instead of modifying the resulting `item` and `includeTree` with `ref` parameters, these values can be optionally overridden by returning an ItemResult with its `Object` and `IncludeTree` properties populated with non-null values.
 
 ## Major Changes